[copy-review] Inject the comment widget into the live email gallery (build.py); the retired CMS screenshot gallery is no longer published, so drop the dead copy from screenshots_spec
